+++ CHANGELOG.md
@@ -25,6 +25,7 @@ Usage:
 * Fix #1194: jkube controller name configuration ignored when using resource fragments
 * Fix #1201: ThorntailV2Generator works with Gradle Plugins
 * Fix #1208: Allow env variables to be passed to a webapp generator s2i builder
+* Fix #1218: Generate yaml files for different environments
 * Fix #1251: Generate a preview of jkube documentation for PR if needed
 * Fix #1259: Add integration test and docs for NameEnricher
 * Fix #1260: Add documentation for PodAnnotationEnricher